requirements.txt
PyPDF2==3.0.1
google-genai>=1.0.0
streamlit==1.32.0
python-dotenv==1.0.1
pytest==8.0.0